pyproject.toml
[project]
name = "nlsh"
version = "0.1.0"
description = "Natural Language Shell - AI-augmented command line interface"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "rich>=13.0.0",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "prompt-toolkit>=3.0.0"
]
readme = "README.md"
requires-python = ">=3.9"

[project.scripts]
nlsh = "nlsh.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 